JS-1597 decouple generated source telemetry modules
